Fix #3824: Move UndefinedBehaviorError into the linker library
