+++ linker-private-library/src/main/scala/scala/scalajs/runtime/UndefinedBehaviorError.scala
